Allow compilation on Windows.
Without this commit, a build on Windows fails with:

differ\diff.go:124 undefined: syscall.Stat_t
