+++ differ/BUILD.bazel
@@ -2,6 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["diff.go"],
+    srcs = [
+        "diff.go",
+        "isatty_darwin.go",
+        "isatty_linux.go",
+        "isatty_windows.go",
+    ],
     visibility = ["//visibility:public"],
 )
