+++ differ/isatty_linux.go
@@ -1,0 +1,30 @@
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+package differ
+
+import "syscall"
+
+// isatty reports whether fd is a tty.
+// Actually it reports whether fd is a character device, which is close enough.
+func isatty(fd int) bool {
+	var st syscall.Stat_t
+	err := syscall.Fstat(fd, &st)
+	if err != nil {
+		return false
+	}
+	return st.Mode&syscall.S_IFMT == syscall.S_IFCHR
+}
